Better but they have some bugs now
I didn't want to merge this with working because the working branch has a working project and these ones have a few bugs. If these end up not working by the deadline, working will still technically work. 
Bugs: Placing pieces for row 1 in and extra column, bot doesn't work as a real bot